deepen: inline SectorId into Onboarding company list, delete sic-sector-map
Replaces 1 shallow module with the SectorId type collocated next to its sole
consumer (the Onboarding company list).

`lib/sec-edgar/sic-sector-map.ts` exported three things:
- `SECTOR_DEFINITIONS` — 0 production callers
- `sicToSector()` — 0 production callers
- `SectorId` type — 1 production caller (`lib/onboarding/popular-companies.ts`)

Only the type alias was actually consumed. The 159-line SIC→GICS lookup table
and lookup function were dead code; deletion test passes (complexity does not
reappear when the module is removed — popular-companies just needs the
SectorId union, which it now declares directly from a 12-line `as const`
array).

Dependency category: 1 (in-process). No adapter needed.

Tests deleted: none (no callers exercised the deleted function).
Tests added: none (the change is type-only at the new interface).

Follows the ADR 0001 pattern (move the registry config into the registry,
delete the side-effect indirection).
